Finished Today Page
The today page is looking pretty good; the tabs are updated to match the mockup, the link to the admin page and linktree work as intended, I've created a ClassCard component which should be useful, as well as a TextBubble component for the tags under a person's name. They're also used to display the location of the class in the ClassCard component. The web version of the app isn't working, so I haven't been able to test on different screen sizes. Hopefully it works on your phone; I will want to try and figure out how to test it at different sizes at some point, because a lot of the organization of the today screen is based on padding and margins that could be potentially janky at different sizes.

Hopefully closes #11