Sponsor & Membership Modals Implemented
I've implement early versions of the Sponsor & Membership Modals. Tapping the membership card button or the SPONSOR PERKS button brings up a modal. The way the images are displayed in both the Sponsor and Membership modals are imperfect, but this is the best implementation I could think of quickly without creating something really complicated, and I wanted to keep the structure simple since I'm adding a lot of code here.